remove old transaction ui components and supporting actions, keep transations reducers they are used for alerts
