--- src/assertions/is-transactions-working.js
@@ -1,4 +1,0 @@
-export default function(isTransactionsWorking) {
-  expect(isTransactionsWorking).toBeDefined();
-  expect(typeof isTransactionsWorking).toEqual("boolean");
-}
