+++ src/modules/markets/actions/load-create-market-history.js
@@ -1,6 +1,5 @@
 import { augur } from "services/augurjs";
 import logError from "utils/log-error";
-import { addMarketCreationTransactions } from "modules/transactions/actions/add-transactions";
 import { loadMarketsInfoIfNotLoaded } from "modules/markets/actions/load-markets-info";
 import { loadUnclaimedFees } from "modules/markets/actions/market-creator-fees-management";
 
@@ -24,7 +23,6 @@ export function loadCreateMarketHistory(options = {}, callback = logError) {
         dispatch(
           loadMarketsInfoIfNotLoaded(marketsCreatedByUser, err => {
             if (err) return callback(err);
-            dispatch(addMarketCreationTransactions(marketsCreatedByUser));
             // TODO save markets created by user to state
             callback(null, marketsCreatedByUser);
           })
