--- src/modules/portfolio/components/transaction-header/transaction-header.styles.less
@@ -1,41 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.TransactionHeader__date {
-  color: @color-white;
-  font-size: @font-size-extra-small;
-  line-height: @font-size-extra-small;
-}
-
-.TransactionHeader__description {
-  font-size: @font-size-large;
-  line-height: @font-size-rather-large;
-  margin-bottom: 0.6rem;
-}
-
-.TransactionHeader__message {
-  &:extend(.caps--large);
-
-  font-weight: 700;
-  margin-bottom: 0.35rem;
-}
-
-.TransactionHeader__message-canceled {
-  color: @color-error;
-  font-size: 0.75rem;
-  letter-spacing: 0;
-  padding-left: 1rem;
-}
-
-.TransactionHeader__status {
-  &:extend(.caps--extra-small);
-
-  color: @color-light-gray;
-  font-weight: 500;
-  margin-bottom: 0.625rem;
-}
-
-@media @breakpoint-mobile-medium {
-  .TransactionHeader__message {
-    line-height: 1.5rem;
-  }
-}
