--- src/modules/portfolio/components/transaction-linked/transaction-linked.styles.less
@@ -1,46 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.LinkedTransaction__message {
-  align-items: center;
-  background-color: @color-dark-grayish-purple;
-  border-top: 1px solid @color-almost-black;
-  color: @color-white;
-  display: flex;
-  justify-content: space-between;
-  padding: 1rem 2rem;
-  text-align: left;
-  width: 100%;
-
-  > div {
-    overflow: hidden;
-    padding-right: 1rem;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-}
-
-.LinkedTransaction__message-text {
-  &:extend(.caps--large);
-
-  font-weight: bold;
-}
-
-.LinkedTransaction__message-canceled {
-  color: @color-error;
-  font-size: 0.75rem;
-  letter-spacing: 0;
-  padding-left: 1rem;
-}
-
-.LinkedTransaction__status {
-  &:extend(.caps--extra-small);
-
-  font-weight: 500;
-  margin-bottom: 0.625rem;
-}
-
-@media @breakpoint-mobile-medium {
-  .LinkedTransaction__message {
-    padding: 1rem;
-  }
-}
